requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
idna==3.3
Pillow==9.1.0
pymongo==3.12.3
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.9
